+++ app/src/controllers/organizations/organization/constants.js
@@ -1,4 +1,4 @@
-/*!
+/*
  * Copyright 2024 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-.organization-projects-container {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
+export const PROJECTS_NAMESPACE = 'projects';
+
+export const ACTIVE_ORGANIZATION_NAMESPACE = 'activeOrganization';
+
+export const FETCH_ORGANIZATION_PROJECTS = 'fetchOrganizationProjects';
