+++ app/src/controllers/organizations/sagas.js
@@ -14,20 +14,12 @@
  * limitations under the License.
  */
 
-import { takeEvery, all, put, select, take } from 'redux-saga/effects';
+import { takeEvery, all, put } from 'redux-saga/effects';
 import { URLS } from 'common/urls';
 import { showDefaultErrorNotification } from 'controllers/notification';
-import { createFetchPredicate, fetchDataAction } from 'controllers/fetch';
-import { organizationsListSelector } from 'controllers/organizations/selectors';
-import { fetchOrganizationsAction } from 'controllers/organizations/actionCreators';
-import { redirect } from 'redux-first-router';
-import { PROJECTS_PAGE } from 'controllers/pages';
-import {
-  FETCH_ORGANIZATIONS,
-  FETCH_ORGANIZATION_PROJECTS,
-  NAMESPACE,
-  ACTIVE_ORGANIZATION_NAMESPACE,
-} from './constants';
+import { fetchDataAction } from 'controllers/fetch';
+import { organizationSagas } from 'controllers/organizations/organization/sagas';
+import { FETCH_ORGANIZATIONS, NAMESPACE } from './constants';
 
 function* fetchOrganizations() {
   try {
@@ -37,38 +29,10 @@ function* fetchOrganizations() {
   }
 }
 
-function* fetchOrganizationProjects({ payload: { organizationSlug } }) {
-  yield put(fetchOrganizationsAction());
-  yield take(createFetchPredicate(NAMESPACE));
-  const organizations = yield select(organizationsListSelector);
-  try {
-    const orgsId = organizations.find((org) => org.slug === organizationSlug)?.id;
-    // TODO: Uncomment this line after implementation of the organizationProjects in backend
-    // yield put(fetchDataAction(PROJECTS_NAMESPACE)(URLS.organizationProjects(orgsId)));
-
-    yield put(fetchDataAction(ACTIVE_ORGANIZATION_NAMESPACE)(URLS.organizationById(orgsId)));
-    yield take(createFetchPredicate(ACTIVE_ORGANIZATION_NAMESPACE));
-
-    if (!orgsId) {
-      throw new Error('Organization not found');
-    }
-  } catch (error) {
-    yield put(
-      redirect({
-        type: PROJECTS_PAGE,
-      }),
-    );
-  }
-}
-
 function* watchFetchOrganizations() {
   yield takeEvery(FETCH_ORGANIZATIONS, fetchOrganizations);
 }
 
-function* watchFetchOrganizationProjects() {
-  yield takeEvery(FETCH_ORGANIZATION_PROJECTS, fetchOrganizationProjects);
-}
-
 export function* organizationsSagas() {
-  yield all([watchFetchOrganizations(), watchFetchOrganizationProjects()]);
+  yield all([watchFetchOrganizations(), organizationSagas()]);
 }
