+++ app/src/controllers/organizations/selectors.js
@@ -14,13 +14,8 @@
  * limitations under the License.
  */
 
-const organizationsSelector = (state) => state.organizations || {};
+export const organizationsSelector = (state) => state.organizations || {};
 
 export const organizationsListSelector = (state) => organizationsSelector(state).list || [];
 
 export const organizationsListLoadingSelector = (state) => organizationsSelector(state).listLoading;
-
-export const activeOrganizationSelector = (state) =>
-  organizationsSelector(state).activeOrganization || {};
-
-export const organizationProjectsSelector = (state) => organizationsSelector(state).projects || [];
