+++ app/src/pages/inside/projectSettingsPageContainer/content/notifications/modals/addEditNotificationModal/addEditNotificationModal.scss
@@ -57,7 +57,7 @@
   }
 }
 
-.dynamicField{
+.dynamicField {
   margin-top: 16px;
   margin-bottom: 16px;
 }
