+++ app/src/pages/organization/organizationProjectsPage/header/projectsPageHeader.scss
@@ -70,7 +70,6 @@
       svg {
         width: 16px;
         height: 16px;
-
       }
     }
   }
